[refactor] home page 내부에서 "use client"; 호출 부분 삭제
[refactor] home page 내용 컴포넌트화 하여 page에서 호출하는 방식으로 "use client"; 호출 부분 삭제
